requirements.txt
asgiref==3.11.0
Django==5.2.10
django-formtools==2.5.1
django-otp==1.7.0
django-phonenumber-field==8.4.0
django-two-factor-auth==1.18.1
Faker==40.1.2
phonenumberslite==8.13.55
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
qrcode==8.2
RapidFuzz==3.14.3
sqlparse==0.5.5
typing_extensions==4.15.0
